Label cache results with paths
When investigating a cahce efficiency drop or a token use surge, it is
necessary to know what paths the cache is missing on.
